+++ lib/Parser.mjs
@@ -1,8 +1,8 @@
 import { Dataset } from './Dataset.mjs';
 import { Tags, LineVal } from './Syntax.mjs';
 
-import { Individual } from './records/Individual.mjs';
-import { Family } from './records/Family.mjs';
+import { Individual } from './structure/Individual.mjs';
+import { Family } from './structure/Family.mjs';
 
 export class Parser {
