+++ lib/index.mjs
@@ -1,8 +1,8 @@
 import { Parser } from './Parser.mjs';
 import { readGedFile } from './helpers/readGedFile.mjs';
 import { printGedcomRecords } from './helpers/printGedcomRecords.mjs';
-import { Structure } from "./Structure.mjs";
-import {Individual} from "./records/Individual.mjs";
+import { Structure } from "./structure/Structure.mjs";
+import {Individual} from "./structure/Individual.mjs";
 
 async function main() {
