+++ lib/structure/Family.mjs
@@ -2,7 +2,7 @@
 // produce CHIL (children). The FAM record may also be used for cultural parallels to this, including nuclear families, marriage, 
 // cohabitation, fostering, adoption, and so on, regardless of the gender of the partners
 
-import { Structure } from "../Structure.mjs";
+import { Structure } from "./Structure.mjs";
 
 export class Family extends Structure  {
     constructor(line) {
