+++ lib/structure/Individual.mjs
@@ -1,7 +1,7 @@
 // The individual record is a compilation of facts or hypothesized facts about an individual.
 // These facts may come from multiple sources.
 
-import { Structure } from "../Structure.mjs";
+import { Structure } from "./Structure.mjs";
 
 export class Individual extends Structure  {
     constructor(line) {
