+++ lib/structure/Structure.mjs
@@ -1,4 +1,4 @@
-import { LineVal } from "./Syntax.mjs";
+import { LineVal } from "../Syntax.mjs";
 
 export class Structure {
     // a structure consists of a level, optional cross-reference identifier, tag and optional line value
